Assets+Gfx: Make Sprite, SpriteGroup, and Ninepatch into Assets
