+++ src/Gfx/AssetLoader.cpp
@@ -8,7 +8,9 @@
 #include <Assets/AssetManager.h>
 #include <Gfx/BitmapFont.h>
 #include <Gfx/Cursor.h>
+#include <Gfx/Ninepatch.h>
 #include <Gfx/Palette.h>
+#include <Gfx/Sprite.h>
 #include <Gfx/TextDocument.h>
 #include <Util/Blob.h>
 #include <Util/ErrorOr.h>
@@ -28,6 +30,11 @@ void AssetLoader::register_types(AssetManager& assets)
     assets.asset_loaders_by_type[AssetType::PaletteDefs] = this;
     assets.asset_loaders_by_type[AssetType::Palette] = this;
 
+    assets.fileExtensionToType.put(assets.assetStrings.intern("sprites"_s), AssetType::SpriteDefs);
+    assets.asset_loaders_by_type[AssetType::SpriteDefs] = this;
+    assets.asset_loaders_by_type[AssetType::Sprite] = this;
+    assets.asset_loaders_by_type[AssetType::Ninepatch] = this;
+
     assets.fileExtensionToType.put(assets.assetStrings.intern("txt"_s), AssetType::TextDocument);
     assets.asset_loaders_by_type[AssetType::TextDocument] = this;
 }
@@ -36,7 +43,9 @@ void AssetLoader::create_placeholder_assets(AssetManager& assets)
 {
     assets.set_placeholder_asset(AssetType::BitmapFont, adopt_own(*new BitmapFont));
     assets.set_placeholder_asset(AssetType::Cursor, adopt_own(*new Cursor));
+    assets.set_placeholder_asset(AssetType::Ninepatch, Ninepatch::make_placeholder());
     assets.set_placeholder_asset(AssetType::Palette, adopt_own(*new Palette));
+    assets.set_placeholder_asset(AssetType::Sprite, SpriteGroup::make_placeholder());
     assets.set_placeholder_asset(AssetType::TextDocument, adopt_own(*new TextDocument));
 }
 
@@ -56,6 +65,9 @@ ErrorOr<NonnullOwnPtr<Asset>> AssetLoader::load_asset(AssetMetadata& metadata, B
 
     if (metadata.type == AssetType::PaletteDefs)
         return to_error_or_asset(Palette::load_defs(metadata, file_data));
+
+    if (metadata.type == AssetType::SpriteDefs)
+        return to_error_or_asset(load_sprite_defs(metadata, file_data));
 
     if (metadata.type == AssetType::TextDocument)
         return to_error_or_asset(TextDocument::load(metadata, file_data));
