+++ src/Gfx/CMakeLists.txt
@@ -4,6 +4,7 @@ target_sources(CitySim PRIVATE
     Camera.cpp
     Colour.cpp
     Cursor.cpp
+    Ninepatch.cpp
     OpenGL.cpp
     Palette.cpp
     RenderBuffer.cpp
