+++ src/Gfx/Forward.h
@@ -11,12 +11,12 @@ struct BitmapFontGlyph;
 class Camera;
 class Colour;
 class Cursor;
-struct Ninepatch;
+class Ninepatch;
 class Palette;
 struct RenderBuffer;
 class Renderer;
 struct Shader;
-struct Sprite;
-struct SpriteGroup;
+class Sprite;
+class SpriteGroup;
 class SpriteRef;
 struct Texture;
