+++ src/Gfx/Renderer.cpp
@@ -8,6 +8,7 @@
 #include <Assets/AssetManager.h>
 #include <Debug/Debug.h>
 #include <Gfx/Cursor.h>
+#include <Gfx/Ninepatch.h>
 #include <Gfx/OpenGL.h>
 #include <Settings/Settings.h>
