+++ src/UI/Drawable.cpp
@@ -6,6 +6,7 @@
 
 #include "Drawable.h"
 #include <Assets/AssetManager.h>
+#include <Gfx/Ninepatch.h>
 #include <UI/UITheme.h>
 
 namespace UI {
@@ -27,7 +28,7 @@ void Drawable::preparePlaceholder(RenderBuffer* buffer)
     } break;
 
     case DrawableType::Ninepatch: {
-        placeholder = appendDrawNinepatchPlaceholder(buffer, dynamic_cast<DeprecatedAsset&>(*style->ninepatch.get().loaded_asset).ninepatch.texture, renderer.shaderIds.textured);
+        placeholder = appendDrawNinepatchPlaceholder(buffer, dynamic_cast<Ninepatch&>(*style->ninepatch.get().loaded_asset).texture, renderer.shaderIds.textured);
     } break;
 
     case DrawableType::Sprite: {
@@ -54,7 +55,7 @@ void Drawable::fillPlaceholder(Rect2I bounds)
     } break;
 
     case DrawableType::Ninepatch: {
-        fillDrawNinepatchPlaceholder(&placeholder.get<DrawNinepatchPlaceholder>(), bounds, &dynamic_cast<DeprecatedAsset&>(*style->ninepatch.get().loaded_asset).ninepatch, style->color);
+        fillDrawNinepatchPlaceholder(&placeholder.get<DrawNinepatchPlaceholder>(), bounds, &dynamic_cast<Ninepatch&>(*style->ninepatch.get().loaded_asset), style->color);
     } break;
 
     case DrawableType::Sprite: {
@@ -82,7 +83,7 @@ void Drawable::draw(RenderBuffer* buffer, Rect2I bounds)
     } break;
 
     case DrawableType::Ninepatch: {
-        drawNinepatch(buffer, bounds, renderer.shaderIds.textured, &dynamic_cast<DeprecatedAsset&>(*style->ninepatch.get().loaded_asset).ninepatch, style->color);
+        drawNinepatch(buffer, bounds, renderer.shaderIds.textured, &dynamic_cast<Ninepatch&>(*style->ninepatch.get().loaded_asset), style->color);
     } break;
 
     case DrawableType::Sprite: {
